@@ -4,7 +4,7 @@
     "type": "library",
     "license": "MIT",
     "require": {
-        "php": "~8.0.0|~8.1.0|~8.2.0|~8.3.0",
+        "php": "~8.1.0|~8.2.0|~8.3.0",
         "ext-dom": "*",
         "ext-bcmath": "*",
         "symfony/validator": "^4.4|^5.0|^6.0|^7.0",
