@@ -1,5 +1,6 @@
 <?php declare(strict_types=1);
 
+use Fpdf\Fpdf;
 use Sprain\SwissQrBill as QrBill;
 use Sprain\SwissQrBill\PaymentPart\Output\DisplayOptions;
 
@@ -10,7 +11,7 @@ require __DIR__ . '/../example.php';
 
 // 2. Create an FPDF instance (or use an existing one from your project)
 // – alternatively, an instance of \setasign\Fpdi\Fpdi() is also accepted by FpdfOutput.
-$fpdf = new \Fpdf\Fpdf('P', 'mm', 'A4');
+$fpdf = new Fpdf('P', 'mm', 'A4');
 
 // In case your server does not support "allow_url_fopen", use this way to create your FPDF instance:
 // $fpdf = new class('P', 'mm', 'A4') extends \Fpdf\Fpdf {
