@@ -16,14 +16,13 @@ final class AlternativeScheme implements QrCodeableInterface, SelfValidatableInt
 {
     use SelfValidatableTrait;
 
-    /**
-     * Parameter character chain of the alternative scheme
-     */
-    private string $parameter;
-
-    private function __construct(string $parameter)
+    private function __construct(
+        /**
+         * Parameter character chain of the alternative scheme
+         */
+        private string $parameter
+    )
     {
-        $this->parameter = $parameter;
     }
 
     public static function create(string $parameter): self
