@@ -16,19 +16,16 @@ final class PaymentAmountInformation implements QrCodeableInterface, SelfValidat
     public const CURRENCY_EUR = 'EUR';
 
     /**
-     * The payment amount due
-     */
-    private ?float $amount;
-
-    /**
      * Payment currency code (ISO 4217)
      */
     private string $currency;
 
-    private function __construct(string $currency, ?float $amount)
+    private function __construct(string $currency, /**
+     * The payment amount due
+     */
+    private ?float $amount)
     {
         $this->currency = strtoupper($currency);
-        $this->amount = $amount;
     }
 
     public static function create(string $currency, ?float $amount = null): self
