@@ -28,17 +28,17 @@ final class FpdiOutputTest extends TestCase
             [
                 'layout' => (new DisplayOptions())->setPrintable(false),
                 'format' => QrCode::FILE_FORMAT_PNG,
-                'file' => dirname(dirname(dirname(__DIR__))) . '/TestData/FpdfOutput/' . $name . '.pdf'
+                'file' => dirname(__DIR__, 3) . '/TestData/FpdfOutput/' . $name . '.pdf'
             ],
             [
                 'layout' => (new DisplayOptions())->setPrintable(true),
                 'format' => QrCode::FILE_FORMAT_PNG,
-                'file' => dirname(dirname(dirname(__DIR__))) . '/TestData/FpdfOutput/' . $name . '.print.pdf'
+                'file' => dirname(__DIR__, 3) . '/TestData/FpdfOutput/' . $name . '.print.pdf'
             ],
             [
                 'layout' => (new DisplayOptions())->setPrintable(false)->setDisplayScissors(true),
                 'format' => QrCode::FILE_FORMAT_PNG,
-                'file' => dirname(dirname(dirname(__DIR__))) . '/TestData/FpdfOutput/' . $name . '.scissors.pdf'
+                'file' => dirname(__DIR__, 3) . '/TestData/FpdfOutput/' . $name . '.scissors.pdf'
             ],
             [
                 'layout' => (new DisplayOptions())->setPrintable(false)->setDisplayScissors(true)->setPositionScissorsAtBottom(true),
@@ -123,11 +123,7 @@ final class FpdiOutputTest extends TestCase
         // Extract actual pdf content and ignore all meta data which may differ in different versions of Fpdf
         $pattern = '/stream(.*?)endstream/s';
         preg_match($pattern, $fileContents, $matches);
-
-        if (isset($matches[1])) {
-            return $matches[1];
-        }
-        return null;
+        return $matches[1] ?? null;
     }
 
     private function instantiateFpdi($withMemImageSupport = null, $withBundledTrait = null): Fpdi
